+++ crates/miden-lib/src/account/auth/mod.rs
@@ -1,6 +1,18 @@
 mod no_auth;
 pub use no_auth::NoAuth;
 
+mod ecdsa_k256_keccak;
+pub use ecdsa_k256_keccak::AuthEcdsaK256Keccak;
+
+mod ecdsa_k256_keccak_acl;
+pub use ecdsa_k256_keccak_acl::{AuthEcdsaK256KeccakAcl, AuthEcdsaK256KeccakAclConfig};
+
+mod ecdsa_k256_keccak_multisig;
+pub use ecdsa_k256_keccak_multisig::{
+    AuthEcdsaK256KeccakMultisig,
+    AuthEcdsaK256KeccakMultisigConfig,
+};
+
 mod rpo_falcon_512;
 pub use rpo_falcon_512::AuthRpoFalcon512;