+++ crates/miden-lib/src/account/interface/mod.rs
@@ -15,6 +15,9 @@ use crate::AuthScheme;
 use crate::account::components::{
     basic_fungible_faucet_library,
     basic_wallet_library,
+    ecdsa_k256_keccak_acl_library,
+    ecdsa_k256_keccak_library,
+    ecdsa_k256_keccak_multisig_library,
     network_fungible_faucet_library,
     no_auth_library,
     rpo_falcon_512_acl_library,
@@ -144,6 +147,19 @@ impl AccountInterface {
                 AccountComponentInterface::NetworkFungibleFaucet(_) => {
                     component_proc_digests.extend(
                         network_fungible_faucet_library().mast_forest().procedure_digests(),
+                    );
+                },
+                AccountComponentInterface::AuthEcdsaK256Keccak(_) => {
+                    component_proc_digests
+                        .extend(ecdsa_k256_keccak_library().mast_forest().procedure_digests());
+                },
+                AccountComponentInterface::AuthEcdsaK256KeccakAcl(_) => {
+                    component_proc_digests
+                        .extend(ecdsa_k256_keccak_acl_library().mast_forest().procedure_digests());
+                },
+                AccountComponentInterface::AuthEcdsaK256KeccakMultisig(_) => {
+                    component_proc_digests.extend(
+                        ecdsa_k256_keccak_multisig_library().mast_forest().procedure_digests(),
                     );
                 },
                 AccountComponentInterface::AuthRpoFalcon512(_) => {