+++ crates/miden-tx/src/host/storage_delta_tracker.rs
@@ -21,6 +21,8 @@ use miden_objects::account::{
 /// that slot.
 #[derive(Debug, Clone)]
 pub struct StorageDeltaTracker {
+    /// Flag indicating whether this delta is for a new account.
+    is_account_new: bool,
     /// The _initial_ storage header of the native account against which the transaction is
     /// executed. This is only used to look up the initial values of storage _value_ slots, while
     /// the map slots are unused.
@@ -48,6 +50,7 @@ impl StorageDeltaTracker {
         };
 
         let mut storage_delta_tracker = Self {
+            is_account_new: account.is_new(),
             storage_header: initial_storage_header,
             init_maps: BTreeMap::new(),
             delta: AccountStorageDelta::new(),
@@ -58,9 +61,9 @@ impl StorageDeltaTracker {
             (0..u8::MAX).zip(account.storage().header().slots()).for_each(
                 |(slot_idx, (slot_type, value))| match slot_type {
                     StorageSlotType::Value => {
-                        // Note that we can insert the value unconditionally as the delta will be
-                        // normalized before the commitment is computed.
-                        storage_delta_tracker.set_item(slot_idx, Word::empty(), *value);
+                        // For new accounts, all values should be added to the delta, even empty
+                        // words, so that the final delta includes the storage slot.
+                        storage_delta_tracker.set_item(slot_idx, *value);
                     },
                     StorageSlotType::Map => {
                         let storage_map = account
@@ -68,6 +71,9 @@ impl StorageDeltaTracker {
                             .maps()
                             .find(|map| map.root() == *value)
                             .expect("storage map should be present in partial storage");
+
+                        // Make sure each map is represented by at least an empty storage map delta.
+                        storage_delta_tracker.delta.insert_empty_map_delta(slot_idx);
 
                         storage_map.entries().for_each(|(key, value)| {
                             storage_delta_tracker.set_map_item(
@@ -89,11 +95,8 @@ impl StorageDeltaTracker {
     // --------------------------------------------------------------------------------------------
 
     /// Updates a value slot.
-    pub fn set_item(&mut self, slot_index: u8, prev_value: Word, new_value: Word) {
-        // Don't update the delta if the new value matches the old one.
-        if prev_value != new_value {
-            self.delta.set_item(slot_index, new_value);
-        }
+    pub fn set_item(&mut self, slot_index: u8, new_value: Word) {
+        self.delta.set_item(slot_index, new_value);
     }
 
     /// Updates a map slot.
@@ -127,22 +130,33 @@ impl StorageDeltaTracker {
     /// - removing entries for map slot updates where for a given key, the new value is equal to the
     ///   initial value at the beginning of transaction execution.
     fn normalize(self) -> AccountStorageDelta {
-        let Self { storage_header, init_maps, delta } = self;
+        let Self {
+            is_account_new,
+            storage_header,
+            init_maps,
+            delta,
+        } = self;
         let (mut value_slots, mut map_slots) = delta.into_parts();
 
-        // Keep only the values whose new value is different from the initial value.
-        value_slots.retain(|slot_idx, new_value| {
-            // SAFETY: The header in the initial storage is the one from the account against which
-            // the transaction is executed, so accessing that slot index should be fine.
-            let (_, initial_value) =
-                storage_header.slot(*slot_idx as usize).expect("index should be in bounds");
-            new_value != initial_value
-        });
+        // Skip normalization of value slots for new accounts. Since the initial value for
+        // normalization defaults to Word::empty, this prevents slots being removed that are validly
+        // created with an empty value.
+        if !is_account_new {
+            // Keep only the values whose new value is different from the initial value.
+            value_slots.retain(|slot_idx, new_value| {
+                // SAFETY: The header in the initial storage is the one from the account against
+                // which the transaction is executed, so accessing that slot index
+                // should be fine.
+                let (_, initial_value) =
+                    storage_header.slot(*slot_idx as usize).expect("index should be in bounds");
+                new_value != initial_value
+            });
+        }
 
         // On the key-value level: Keep only the key-value pairs whose new value is different from
         // the initial value.
         // On the map level: Keep only the maps that are non-empty after its key-value pairs have
-        // been normalized.
+        // been normalized, or if the account is new.
         map_slots.retain(|slot_idx, map_delta| {
             let init_map = init_maps.get(slot_idx);
 
@@ -155,8 +169,9 @@ impl StorageDeltaTracker {
                 });
             }
 
-            // Only retain the map delta if it still contains values after normalization.
-            !map_delta.is_empty()
+            // Only retain the map delta if the account is new or if it still contains values after
+            // normalization.
+            self.is_account_new || !map_delta.is_empty()
         });
 
         AccountStorageDelta::from_parts(value_slots, map_slots)